URL Validation

+++ matchmaking/src/main/java/org/eclipse/tractusx/agents/service/DataManagement.java
@@ -42,6 +42,7 @@ import org.eclipse.tractusx.agents.utils.TypeManager;
 
 import java.io.IOException;
 import java.net.URLEncoder;
+import java.net.URI;
 import java.nio.charset.StandardCharsets;
 import java.util.List;
 import java.util.UUID;
@@ -426,6 +427,26 @@ public class DataManagement {
     }
 
     /**
+     * URL Validation
+     *
+     * @param url URL string for validation
+     * @return Returns true if the URL is valid, false otherwise.
+     */
+    public static boolean isValid(String url) 
+    { 
+        // Try creating a valid URL
+        try { 
+            new URI(url).toURL(); 
+            return true; 
+        } 
+          
+        // If there was an Exception while creating URL object 
+        catch (Exception e) { 
+            return false; 
+        } 
+    } 
+
+    /**
      * deletes an existing aseet
      *
      * @param assetId key of the asset
@@ -435,19 +456,24 @@ public class DataManagement {
     public IdResponse deleteAsset(String assetId) throws IOException {
         String version = "/v3";
         var url = String.format(ASSET_UPDATE_CALL, config.getControlPlaneManagementProviderUrl(), version, assetId);
-        var request = new Request.Builder().url(url).delete();
-        config.getControlPlaneManagementHeaders().forEach(request::addHeader);
-        try (var response = httpClient.newCall(request.build()).execute()) {
-            ResponseBody body = response.body();
-            if (response.isSuccessful() && body != null) {
-                return JsonLd.processIdResponse(body.string());
-            } else {
-                monitor.warning(format("Failure in calling the control plane at %s. Ignoring", url));
-                return null;
+        if(isValid(url)){
+            var request = new Request.Builder().url(url).delete();
+            config.getControlPlaneManagementHeaders().forEach(request::addHeader);
+            try (var response = httpClient.newCall(request.build()).execute()) {
+                ResponseBody body = response.body();
+                if (response.isSuccessful() && body != null) {
+                    return JsonLd.processIdResponse(body.string());
+                } else {
+                    monitor.warning(format("Failure in calling the control plane at %s. Ignoring", url));
+                    return null;
+                }
+            } catch (Exception e) {
+                monitor.severe(format("Error in calling the control plane at %s", url), e);
+                throw e;
             }
-        } catch (Exception e) {
-            monitor.severe(format("Error in calling the control plane at %s", url), e);
-            throw e;
+        } else {
+            monitor.warning(format("Invalid URL", url));
+            return null;
         }
     }
